requirements.txt
yt-dlp
ytmusicapi
mutagen
requests
pillow